Updated File Name and removed delay
I updated the file names for both the username and password files for hydra to grab from. I did this in order to make it even more user friendly. I also removed the time delay as I’m going to create a separate command for individuals who want to automate at the event of WiFi connection compared to those who want to trigger it with User Interaction.